+++ src/main/resources/application.properties
@@ -30,6 +30,8 @@ engagement.api/mp-rest/url=${ENGAGEMENT_API_URL:http://git-api:8080}
 mp.openapi.extensions.smallrye.info.title=LodeStar Artifacts API
 quarkus.swagger-ui.theme=muted
 
+%test.quarkus.scheduler.enabled=false
+
 gitlab.personal.access.token=${GITLAB_TOKEN:t}
 
 # Application
